Rename components/layout.tsx to components/navigation.tsx
