+++ src/app/apartments/[slug]/page.tsx
@@ -1,4 +1,4 @@
-import { Bottombar } from "@/app/components/layout";
+import { Bottombar } from "@/app/components/navigation";
 import { getApartment, updateRoom, navigateHome } from "./actions";
 import Image from "next/image";
 
@@ -7,7 +7,7 @@ export default async function Apartment({ params }: { params: { slug: string } }
 
   return (
     <>
-      <main className="h-screen flex flex-col items-center py-6 px-12">
+      <main className="flex flex-col grow items-center py-6 px-12">
         <div className="text-center mb-3">
           <h1 className="text-3xl">{apartment.name}</h1>
           <h2 className="text-xl">{apartment.location}</h2>
